Add automatic flow.
- When the script is called without arguments it will attempt to detect
if the current kernel is already custom. If it isn't it will go through
the compile process followed by migration.

If it is, it will give the user the option to migrate. If they choose
yes it will go through the migrate flow. If they choose no, it will give
them the option to re-compile anyway or abort.