Rewrite and improve Nvidia's zram systemd script
Improve various facets of the stock script that comes with the Jetson
(even the latest as of 19/4/20) such as

- Being able to run this more than once so you can change the zram device block sizes on the fly.
- Improved swap integration such as enabling discards for the zram swap
- Removing the hard-coded math to determine block size (it will always
use 2GB / 512MB per core), currently setting to use 1GB per core but
doing this with a divisor variable that will come from separate config
later.
- Attempting to improve the compression algorithm to something that is
faster and/or better at compressing (higher ratio of compression).