Support custom (de)compression algorithm
- Add support for a custom (de)compression algorithm rather than only
opting for the default lzo. I've added notes around alternatives.

I've got to do some research and find out the best algorithm for this
device and use-case and then add the support to the kernel and set it
here and in the external config file (that I'm yet to support).